\name{Core}
\alias{Core}
\title{
Internal function of JAMP to generate folders and logs}
\description{
You are not intendet to see this ; )
}
\usage{
Core()
}
\arguments{
  \item{workingonit}{Imput fasta file.}
}
\details{
nope}
\value{
}
\references{
}
\author{
}
\note{
}


\seealso{
\code{\link{NA}}
}
\examples{}